// src/service/Product/Product.type.ts
export interface TProductAPI {
    maPhim: number;
    tenPhim: string;
    biDanh: string;
    trailer: string;
    hinhAnh: string;
    moTa: string;
    maNhom: string;
    ngayKhoiChieu: string;
    danhGia: number;
    hot: boolean;
    dangChieu: boolean;
    sapChieu: boolean;
  }

  export interface TProduct{
    maPhim: number;
    tenPhim: string;
    trailer: string;
    hinhAnh: string;
    ngayKhoiChieu: string;
    hot: boolean;
    dangChieu: boolean;
    sapChieu: boolean;
  }